Small after-rebase fixes
