+++ ivy_tests/test_ivy/test_functional/test_core/test_dtype.py
@@ -1,8 +1,5 @@
 """Collection of tests for unified dtype functions."""
 
-from functools import reduce  # for making strategy
-from operator import mul  # for making strategy
-from typing import Tuple
 
 # global
 import numpy as np
@@ -16,7 +13,10 @@ import ivy.functional.backends.numpy as ivy_np
 import ivy.functional.backends.tensorflow as ivy_tf
 import ivy.functional.backends.torch as ivy_torch
 import ivy_tests.test_ivy.helpers as helpers
-from ivy_tests.test_ivy.helpers import handle_cmd_line_args
+from ivy_tests.test_ivy.helpers import (
+    handle_cmd_line_args,
+    array_and_broadcastable_shape,
+)
 
 
 # dtype objects
@@ -34,88 +34,6 @@ def test_dtype_instances(device):
     assert ivy.exists(ivy.float32)
     assert ivy.exists(ivy.float64)
     assert ivy.exists(ivy.bool)
-
-
-# Functions to use in strategy #
-# ---------------------------- #
-
-# from array-api repo
-def _broadcast_shapes(shape1, shape2):
-    """Broadcasts `shape1` and `shape2`"""
-    N1 = len(shape1)
-    N2 = len(shape2)
-    N = max(N1, N2)
-    shape = [None for _ in range(N)]
-    i = N - 1
-    while i >= 0:
-        n1 = N1 - N + i
-        if N1 - N + i >= 0:
-            d1 = shape1[n1]
-        else:
-            d1 = 1
-        n2 = N2 - N + i
-        if N2 - N + i >= 0:
-            d2 = shape2[n2]
-        else:
-            d2 = 1
-
-        if d1 == 1:
-            shape[i] = d2
-        elif d2 == 1:
-            shape[i] = d1
-        elif d1 == d2:
-            shape[i] = d1
-        else:
-            raise Exception("Broadcast error")
-
-        i = i - 1
-
-    return tuple(shape)
-
-
-# from array-api repo
-def broadcast_shapes(*shapes):
-    if len(shapes) == 0:
-        raise ValueError("shapes=[] must be non-empty")
-    elif len(shapes) == 1:
-        return shapes[0]
-    result = _broadcast_shapes(shapes[0], shapes[1])
-    for i in range(2, len(shapes)):
-        result = _broadcast_shapes(result, shapes[i])
-    return result
-
-
-# np.prod and others have overflow and math.prod is Python 3.8+ only
-def prod(seq):
-    return reduce(mul, seq, 1)
-
-
-# from array-api repo
-def mutually_broadcastable_shapes(
-    num_shapes: int,
-    *,
-    base_shape: Tuple[int, ...] = (),
-    min_dims: int = 1,
-    max_dims: int = 4,
-    min_side: int = 1,
-    max_side: int = 4,
-):
-    if max_dims is None:
-        max_dims = min(max(len(base_shape), min_dims) + 5, 32)
-    if max_side is None:
-        max_side = max(base_shape[-max_dims:] + (min_side,)) + 5
-    return (
-        helpers.nph.mutually_broadcastable_shapes(
-            num_shapes=num_shapes,
-            base_shape=base_shape,
-            min_dims=min_dims,
-            max_dims=max_dims,
-            min_side=min_side,
-            max_side=max_side,
-        )
-        .map(lambda BS: BS.input_shapes)
-        .filter(lambda shapes: all(prod(i for i in s if i > 0) < 1000 for s in shapes))
-    )
 
 
 # for data generation in multiple tests
@@ -185,7 +103,7 @@ def test_astype(
 @st.composite
 def broadcastable_arrays(draw, dtypes):
     num_arrays = st.shared(helpers.ints(min_value=2, max_value=5), key="num_arrays")
-    shapes = draw(num_arrays.flatmap(mutually_broadcastable_shapes))
+    shapes = draw(num_arrays.flatmap(helpers.mutually_broadcastable_shapes))
     dtypes = draw(dtypes)
     arrays = []
     for c, (shape, dtype) in enumerate(zip(shapes, dtypes), 1):
@@ -238,7 +156,6 @@ def array_and_broadcastable_shape(draw, dtype):
         label="shape",
     )
     return x, to_shape
-
 
 @handle_cmd_line_args
 @given(
